requirements.txt
Django==3.1.12
djangorestframework==3.12.4
djongo==1.3.7
sqlparse==0.2.4
pandas
requests
openpyxl
psycopg2-binary